move fru editor element into new module
