+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "editor.fontFamily": "JetBrains Mono, Consolas, 'Courier New', monospace"
+}
