+++ fru_gen.toml
@@ -1,0 +1,33 @@
+
+[common]
+file_size = 256
+
+internal_area = "Enabled"
+chassis_area = "Enabled"
+board_area = "Enabled"
+product_area = "Enabled"
+
+
+[chassis]
+type = "Rack Mount Chassis"
+part_number = "CHS1234"
+serial_number = "SN5678"
+extra = "Chassis extra"
+
+[board]
+manufacturer = "qwfqwfg"
+product_name = "Board124"
+serial_number = "SN12345"
+part_number = "BP9876"
+fru_file_id = "FRU123"
+extra = "Board extra"
+
+[product]
+manufacturer = "ProductMFC"
+product_name = "Product1"
+part_number = "PN5678"
+version = "V1.0.0"
+serial_number = "SN123456"
+asset_tag = "AssetTag"
+extra = "Product extra"
+
