+++ output.yaml
@@ -1,0 +1,17 @@
+Chassis_type: ""
+Chassis_Part_Number: "123456"
+Chassis_Serial_Number: "456879"
+Chassis_Extra: "456"
+Board_Manufacturer: "1231"
+Board_Product_Name: "89156"
+Board_Serial_Number: "91"
+Board_Part_Number: "234"
+Board_Fruid: "561"
+Board_Extra: "261"
+Product_Manufacturer: "59"
+Product_Name: "4198"
+Product_Part_Number: "5614"
+Product_Version: "69146"
+Product_Serial_Number: "3516"
+Product_Asset_Tag: "31"
+Product_Extra: "6351036101056"
